requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
black==24.4.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
fastapi==0.110.1
filelock==3.13.4
flake8==7.0.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
identify==2.5.35
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.8.0
numpy==1.26.4
orjson==3.10.1
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pre-commit==3.7.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycodestyle==2.11.1
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.2
pyright==1.1.348
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.3
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
